requirements.txt
kafka-python==2.9.0
python-dotenv==0.21.0